adicionando novas classes

+++ src/br/com/alura/minhasmusicas/modelos/Audio.java
@@ -4,7 +4,7 @@ public class Audio {
     private String titulo;
     private int totalReproducoes;
     private int totalCurtidas;
-    private double classificacao;
+    private int classificacao;
 
     public String getTitulo() {
         return titulo;
@@ -30,7 +30,7 @@ public class Audio {
         return classificacao;
     }
 
-    public void setClassificacao(double classificacao) {
+    public void setClassificacao(int classificacao) {
         this.classificacao = classificacao;
     }
